fix(SD-LEO-FIX-CREDENTIAL-GUARD-INVERSION-001): the unit tier's credential guard applied only when it was not needed
tests/setup.unit.js applied its synthetic sentinel with ||=, so it took effect
exactly when nothing was set and was skipped exactly when a real Supabase URL
was present. fr-c-generator.test.js then gated INSERT/UPDATE/DELETE against four
production tables on skipIf(!HAS_REAL_DB) -- a predicate satisfied by exactly
the condition the guard existed to prevent.

Measured, not inferred: 11 residual production strategic_directives_v2 rows
carry that suite's fc000000- fixture venture_id, 2026-05-04 to 2026-07-07 across
8 dates, one still non-terminal in the live belt. A lower bound -- cleanup added
later means clean runs leave no residue.

The SD framed CI as the safe side. It is the primary harm path: test-coverage.yml
exported real secrets and ran the whole unit project on every push to main, with
continue-on-error hiding the red but not the writes. Both injection sites are
removed; their hang rationale cited tests/setup.js and the smoke suite, neither
of which the unit tier loads.

FR-2 became a post-condition rather than the ambient check the SD specified.
Built the specified way first and measured: it aborted an ordinary local run,
because vitest.config.js loads .env into the parent and every fork inherits it.
Ambient credentials are the normal case here, and a fence that fires when
nothing is wrong is deleted by the first person it stops.

The live-write block moved to fr-c-generator.db.test.js rather than being
re-gated in place. The repo's own DB-test guard made the point: once the
sentinel is unconditional, a live-DB suite in a unit-project path can never run,
so it was misrouted and not merely mis-gated.

Both mutations verified to kill: reverting to ||= trips the fence naming this
machine's live production ref; moving the fence above the assignment fails the
spawn test.


Fleet-Worker: Alpha-5
