+++ .github/workflows/test-coverage.yml
@@ -67,16 +67,27 @@ jobs:
         continue-on-error: true
         env:
           NODE_OPTIONS: --experimental-vm-modules
-          # SD-LEO-INFRA-VITEST-EXITS-61S-001: tests/setup.js falls back to
-          # https://test.invalid.local when SUPABASE_URL is unset. Smoke tests
-          # then issue real .from('table').select() calls that hang in supabase-js
-          # fetch retry against the unresolvable host until vitest's testTimeout
-          # (60s) force-kills the worker — which exits 1 with no test failure.
-          # Pass the secrets so DB-touching tests have a real endpoint.
-          SUPABASE_URL: ${{ secrets.SUPABASE_URL }}
-          SUPABASE_SERVICE_ROLE_KEY: ${{ secrets.SUPABASE_SERVICE_ROLE_KEY }}
-          NEXT_PUBLIC_SUPABASE_URL: ${{ secrets.SUPABASE_URL }}
-          NEXT_PUBLIC_SUPABASE_ANON_KEY: ${{ secrets.NEXT_PUBLIC_SUPABASE_ANON_KEY }}
+          # SD-LEO-FIX-CREDENTIAL-GUARD-INVERSION-001 (FR-5): the real Supabase secrets that used
+          # to be set here are REMOVED, and this was the most dangerous of the two injection sites.
+          #
+          # This step runs `npm run test:coverage` = `vitest run --project unit --coverage` — the
+          # WHOLE unit project, unscoped — on every PR touching src/scripts/lib/server.js and on
+          # every push to main and develop. tests/unit/lib/eva/quality-findings/fr-c-generator.test.js
+          # is collected by that project and gated its INSERT/UPDATE/DELETE block against four
+          # production tables on `skipIf(!HAS_REAL_DB)` — a predicate satisfied by exactly these
+          # secrets. So this workflow was the PRIMARY path by which unit tests wrote to production;
+          # `continue-on-error: true` below hid the resulting red but never the writes. 11 rows
+          # carrying that suite's fixture sentinel reached production over two months.
+          #
+          # WHY THE OLD RATIONALE NO LONGER APPLIES. It cited SD-LEO-INFRA-VITEST-EXITS-61S-001:
+          # tests/setup.js falls back to an unresolvable host, smoke tests then issue real .from()
+          # calls that hang until the 60s testTimeout kills the worker. That names tests/setup.js
+          # and the SMOKE suite — but the unit project loads tests/setup.unit.js, and the smoke
+          # suite is EXCLUDED from it (SMOKE_INCLUDE in vitest.config.js) and runs in its own
+          # project against setup.db.js. The hang this guarded against cannot occur in this tier.
+          #
+          # MEASURED, not assumed: with the sentinel forced and no credentials, tests/unit/eva/
+          # stage-zero/ runs 1736 tests green in 13.01s with zero hangs.
         # QF-20260426-465: vitest dropped CLI --json in 1.x->3.x; use --reporter=json instead.
         run: npm run test:coverage -- --reporter=json --outputFile=test-results.json
